requirements.txt
alpaca_trade_api
psycopg2-binary
asyncpg
aiohttp
requests
wheel